analysis: move some functions to lib.rs
